Add files via upload
added the engine for anti chess.
anti chess is like normal chess when it comes to how the pieces move but there is one main difference. In anti chess, you win by losing all your pieces. An extra rule is that if you have the choice of eating a piece from your opponent, you have to do so.